+++ src/Serialize.js
@@ -28,6 +28,8 @@ export function stringify (value) {
 
 	if (identity < Enum.text) {
 		switch (identity) {
+			case Enum.portal:
+				return '<' + Enum.offscreen + '>' + stringify(children[0]) + '</' + Enum.offscreen + '>'
 			case Enum.component:
 				return stringify(children[0])
 			case Enum.element: